// bff/src/grpc/gen/candle/wishlist/v1/wishlist.ts
export interface AddWishlistItemRequest {
  userId: string;
  symbol: string;
  displayName: string;
  market: string;
  commandMetadata?: CommandMetadata | undefined;
}

export interface AddWishlistItemResponse {
  item?: WishlistItem | undefined;
}

export interface RemoveWishlistItemRequest {
  userId: string;
  symbol: string;
  commandMetadata?: CommandMetadata | undefined;
}

export interface RemoveWishlistItemResponse {
}

export interface ListWishlistItemsRequest {
  userId: string;
  pageRequest?: PageRequest | undefined;
}
// ...
export const AddWishlistItemRequest: MessageFns<AddWishlistItemRequest> = {
  encode(message: AddWishlistItemRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== "") {
      writer.uint32(10).string(message.userId);
    }
    if (message.symbol !== "") {
      writer.uint32(18).string(message.symbol);
    }
    if (message.displayName !== "") {
      writer.uint32(26).string(message.displayName);
    }
    if (message.market !== "") {
      writer.uint32(34).string(message.market);
    }
    if (message.commandMetadata !== undefined) {
      CommandMetadata.encode(message.commandMetadata, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AddWishlistItemRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddWishlistItemRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.symbol = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.displayName = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.market = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.commandMetadata = CommandMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AddWishlistItemRequest {
    return {
      userId: isSet(object.userId)
        ? globalThis.String(object.userId)
        : isSet(object.user_id)
        ? globalThis.String(object.user_id)
        : "",
      symbol: isSet(object.symbol) ? globalThis.String(object.symbol) : "",
      displayName: isSet(object.displayName)
        ? globalThis.String(object.displayName)
        : isSet(object.display_name)
        ? globalThis.String(object.display_name)
        : "",
      market: isSet(object.market) ? globalThis.String(object.market) : "",
      commandMetadata: isSet(object.commandMetadata)
        ? CommandMetadata.fromJSON(object.commandMetadata)
        : isSet(object.command_metadata)
        ? CommandMetadata.fromJSON(object.command_metadata)
        : undefined,
    };
  },

  toJSON(message: AddWishlistItemRequest): unknown {
    const obj: any = {};
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.symbol !== "") {
      obj.symbol = message.symbol;
    }
    if (message.displayName !== "") {
      obj.displayName = message.displayName;
    }
    if (message.market !== "") {
      obj.market = message.market;
    }
    if (message.commandMetadata !== undefined) {
      obj.commandMetadata = CommandMetadata.toJSON(message.commandMetadata);
    }
    return obj;
  },

  create(base?: DeepPartial<AddWishlistItemRequest>): AddWishlistItemRequest {
    return AddWishlistItemRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AddWishlistItemRequest>): AddWishlistItemRequest {
    const message = createBaseAddWishlistItemRequest();
    message.userId = object.userId ?? "";
    message.symbol = object.symbol ?? "";
    message.displayName = object.displayName ?? "";
    message.market = object.market ?? "";
    message.commandMetadata = (object.commandMetadata !== undefined && object.commandMetadata !== null)
      ? CommandMetadata.fromPartial(object.commandMetadata)
      : undefined;
    return message;
  },
};

function createBaseAddWishlistItemResponse(): AddWishlistItemResponse {
  return { item: undefined };
}

export const AddWishlistItemResponse: MessageFns<AddWishlistItemResponse> = {
  encode(message: AddWishlistItemResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.item !== undefined) {
      WishlistItem.encode(message.item, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AddWishlistItemResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddWishlistItemResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.item = WishlistItem.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AddWishlistItemResponse {
    return { item: isSet(object.item) ? WishlistItem.fromJSON(object.item) : undefined };
  },

  toJSON(message: AddWishlistItemResponse): unknown {
    const obj: any = {};
    if (message.item !== undefined) {
      obj.item = WishlistItem.toJSON(message.item);
    }
    return obj;
  },

  create(base?: DeepPartial<AddWishlistItemResponse>): AddWishlistItemResponse {
    return AddWishlistItemResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AddWishlistItemResponse>): AddWishlistItemResponse {
    const message = createBaseAddWishlistItemResponse();
    message.item = (object.item !== undefined && object.item !== null)
      ? WishlistItem.fromPartial(object.item)
      : undefined;
    return message;
  },
};

function createBaseRemoveWishlistItemRequest(): RemoveWishlistItemRequest {
  return { userId: "", symbol: "", commandMetadata: undefined };
}

export const RemoveWishlistItemRequest: MessageFns<RemoveWishlistItemRequest> = {
  encode(message: RemoveWishlistItemRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== "") {
      writer.uint32(10).string(message.userId);
    }
    if (message.symbol !== "") {
      writer.uint32(18).string(message.symbol);
    }
    if (message.commandMetadata !== undefined) {
      CommandMetadata.encode(message.commandMetadata, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RemoveWishlistItemRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRemoveWishlistItemRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.symbol = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.commandMetadata = CommandMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RemoveWishlistItemRequest {
    return {
      userId: isSet(object.userId)
        ? globalThis.String(object.userId)
        : isSet(object.user_id)
        ? globalThis.String(object.user_id)
        : "",
      symbol: isSet(object.symbol) ? globalThis.String(object.symbol) : "",
      commandMetadata: isSet(object.commandMetadata)
        ? CommandMetadata.fromJSON(object.commandMetadata)
        : isSet(object.command_metadata)
        ? CommandMetadata.fromJSON(object.command_metadata)
        : undefined,
    };
  },

  toJSON(message: RemoveWishlistItemRequest): unknown {
    const obj: any = {};
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.symbol !== "") {
      obj.symbol = message.symbol;
    }
    if (message.commandMetadata !== undefined) {
      obj.commandMetadata = CommandMetadata.toJSON(message.commandMetadata);
    }
    return obj;
  },

  create(base?: DeepPartial<RemoveWishlistItemRequest>): RemoveWishlistItemRequest {
    return RemoveWishlistItemRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<RemoveWishlistItemRequest>): RemoveWishlistItemRequest {
    const message = createBaseRemoveWishlistItemRequest();
    message.userId = object.userId ?? "";
    message.symbol = object.symbol ?? "";
    message.commandMetadata = (object.commandMetadata !== undefined && object.commandMetadata !== null)
      ? CommandMetadata.fromPartial(object.commandMetadata)
      : undefined;
    return message;
  },
};

function createBaseRemoveWishlistItemResponse(): RemoveWishlistItemResponse {
  return {};
}

export const RemoveWishlistItemResponse: MessageFns<RemoveWishlistItemResponse> = {
  encode(_: RemoveWishlistItemResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RemoveWishlistItemResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRemoveWishlistItemResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): RemoveWishlistItemResponse {
    return {};
  },

  toJSON(_: RemoveWishlistItemResponse): unknown {
    const obj: any = {};
    return obj;
  },

  create(base?: DeepPartial<RemoveWishlistItemResponse>): RemoveWishlistItemResponse {
    return RemoveWishlistItemResponse.fromPartial(base ?? {});
  },
  fromPartial(_: DeepPartial<RemoveWishlistItemResponse>): RemoveWishlistItemResponse {
    const message = createBaseRemoveWishlistItemResponse();
    return message;
  },
};
// ...
export type WishlistServiceDefinition = typeof WishlistServiceDefinition;
export const WishlistServiceDefinition = {
  name: "WishlistService",
  fullName: "candle.wishlist.v1.WishlistService",
  methods: {
    addWishlistItem: {
      name: "AddWishlistItem",
      requestType: AddWishlistItemRequest as typeof AddWishlistItemRequest,
      requestStream: false,
      responseType: AddWishlistItemResponse as typeof AddWishlistItemResponse,
      responseStream: false,
      options: {},
    },
    removeWishlistItem: {
      name: "RemoveWishlistItem",
      requestType: RemoveWishlistItemRequest as typeof RemoveWishlistItemRequest,
      requestStream: false,
      responseType: RemoveWishlistItemResponse as typeof RemoveWishlistItemResponse,
      responseStream: false,
      options: {},
    },
    listWishlistItems: {
      name: "ListWishlistItems",
      requestType: ListWishlistItemsRequest as typeof ListWishlistItemsRequest,
      requestStream: false,
      responseType: ListWishlistItemsResponse as typeof ListWishlistItemsResponse,
      responseStream: false,
      options: {},
    },
  },
} as const;